Estilos y demas

+++ database/migrations/2025_09_22_000528_create_inventario_table.php
@@ -17,7 +17,7 @@ return new class extends Migration
                   ->comment('identificador del artículo');
             $table->string('nombre_articulo', 50)
                   ->comment('nombre del producto o artículo');
-            $table->unsignedTinyInteger('cantidad_total')
+            $table->unsignedSmallInteger('cantidad_total')
                   ->comment('cantidad total disponible');
             $table->string('descripcion', 100)
                   ->nullable()
